unifdefall: portability: avoid mktemp -t
GNU mktemp interprets the argument after -t as the entire template
and errors out if it contains no XXXX substring. BSD systems, on
the other hand, treat the argument as a prefix for the mktemp
template and use it verbatim, resulting in long, ugly filenames
like foo.XXXXX.e30GuhHVzU if the string contains embedded Xs.

So avoid -t and use TMPDIR explicitly.

Submitted-by: Jonathan Nieder <jrnieder@gmail.com>