unifdefall: look for unifdef in $(dirname $0)
For debugging and for the test suite it is convenient is
unifdefall can be run in place without relying on unifdef being
installed elsewhere.  So change unifdefall to look for unifdef in
the containing directory.

For compatibility, if unifdef is not present in the directory
containing unifdefall, fall back to searching the $PATH for it.
Some one might have installed unifdef and unifdefall to
different directories.

With this change, 'make test' no longer fails in if1-a.sh when
unifdef is not present on the $PATH.  Also, the test suite will
be a little better at catching new regressions if an old version
of unifdef is installed, since the unifdefall tests will no
longer be testing the installed unifdef.

Submitted-by: Jonathan Nieder <jrnieder@gmail.com>