unifdefall: use cc -E -dM instead of cpp -dM
This is a workaround for Mac OS X, in which ccp invokes gcc
with the -traditional-cpp option which disables C99 variadic
macros. Unfortunately stdio.h requires variadic macro support
so real programs, as well as the tests, are likely to break.

I am not sure if this is the right fix. It is marginally closer to
POSIX, since it has c99 -E but not cpp. But POSIX also lacks cc and
the -dM option.