Fix uninitialized variable in evaluator.
Inner expressions typically return LT_IF without setting their
value pointer when they encounter an unknown value. Binary
operators then blindly use these inner values before checking if
they are unknown - before checking for LT_IF. This is undefined
behaviour, but we usually dodge the bullet - except division can
cause things to explode and expose the bug.

Initializing the value to zero avoids undefined behaviour. But this
means that instead of avoiding division when the denominator is
definitely zero (because it might be unknown and zero) we now only
divide if it is definitely non-zero.