test: try to work better on MinGW
A lot of the test failures are due to CRLFs in stderr which do not
match the bare LFs in the experr files. So, unilaterally strip CRs
from stderr output.

We don't need to do this for stdout, since unifdef writes that in
binary mode so it can match the newline style of the input file.

We might also need to fix the rc files - let's see how this change
works first.