unifdef: more fixes for weird whitespace
The previous attempt (in commit 802c5cd6) to deal with some whitespace
corner cases had a bug when parsing a dangling C comment; rather than
using the general purpose lexer to scan the dangling line, it assumed
the line state must be dirty - which is not necessarily the case for
dangling C comments as it is for long linear whitespace or C++
comments.

The original example that exposed this problem was from Linux
__BACKPORT_LINUX_REGULATOR_DRIVER_H_; the test case in this
commit is based on analysis of the bug.

Reported-By: Pavel.Aronsky@celeno.com
Reported-By: Aviv.Albagli@celeno.com